chore: changes in `changelog.md`
