+++ package.json
@@ -20,37 +20,37 @@
   },
   "dependencies": {
     "@ionic/vue": "8.4.2",
-    "@ionic/vue-router": "8.4.2",
+    "@ionic/vue-router": "8.6.0",
     "@zxing/browser": "^0.1.5",
-    "core-js": "^3.41.0",
-    "ethers": "^6.13.5",
+    "core-js": "^3.42.0",
+    "ethers": "^6.14.3",
     "p-queue": "^8.1.0",
-    "vue": "^3.5.13",
-    "vue-router": "^4.5.0"
+    "vue": "^3.5.16",
+    "vue-router": "^4.5.1"
   },
   "devDependencies": {
-    "@crxjs/vite-plugin": "2.0.0-beta.30",
-    "@types/bun": "^1.2.13",
+    "@crxjs/vite-plugin": "2.0.0-beta.33",
+    "@types/bun": "^1.2.15",
     "@types/archiver": "^6.0.3",
-    "@types/chrome": "^0.0.308",
+    "@types/chrome": "^0.0.326",
     "@types/jest": "^29.5.14",
-    "@typescript-eslint/eslint-plugin": "^8.26.0",
-    "@typescript-eslint/parser": "^8.26.0",
-    "@vitejs/plugin-vue": "^5.2.1",
+    "@typescript-eslint/eslint-plugin": "^8.33.1",
+    "@typescript-eslint/parser": "^8.33.1",
+    "@vitejs/plugin-vue": "^5.2.4",
     "@vue/eslint-config-typescript": "^14.5.0",
     "archiver": "^7.0.1",
-    "eslint": "^9.21.0",
-    "eslint-plugin-vue": "^10.0.0",
+    "eslint": "^9.28.0",
+    "eslint-plugin-vue": "^10.1.0",
     "http-browserify": "^1.7.0",
     "https-browserify": "^1.0.0",
     "jest": "^29.7.0",
-    "sass": "^1.85.1",
+    "sass": "^1.89.1",
     "stream-browserify": "^3.0.0",
-    "ts-jest": "^29.2.6",
-    "typescript": "^5.8.2",
+    "ts-jest": "^29.3.4",
+    "typescript": "^5.8.3",
     "util": "^0.12.5",
-    "vite": "^6.2.0",
-    "vue-tsc": "^2.2.8"
+    "vite": "^6.3.5",
+    "vue-tsc": "^2.2.10"
   },
   "disabledNativeDependencies": {
     "@capacitor/app": "^5.0.6",
