+++ src/extension/manifest.json
@@ -3,8 +3,8 @@
     "name": "__MSG_appName__",
     "description": "__MSG_appDesc__",
     "default_locale": "en",
-    "version": "1.4.20",
-    "version_name": "1.4.20",
+    "version": "1.4.21",
+    "version_name": "1.4.21",
     "icons": {
       "16": "assets/extension-icon/wallet_16.png",
       "32": "assets/extension-icon/wallet_32.png",
