+++ src/views/RequestNetwork.vue
@@ -84,7 +84,6 @@
         :is-open="loading"
         cssClass="my-custom-class"
         message="Please wait..."
-        :duration="4000"
         :key="`k${loading}`"
         @didDismiss="loading = false"
       />
