+++ src/views/SavedReadWriteActionList.vue
@@ -45,7 +45,6 @@
         :is-open="loading"
         cssClass="my-custom-class"
         message="Please wait..."
-        :duration="4000"
         :key="`k${loading}`"
         @didDismiss="loading = false"
       >
